feat: Include Azure as an additional service provider option.
